+++ cbs/cbs.go
@@ -26,11 +26,13 @@ package cbs
 
 import (
 	"context"
+	"fmt"
 	"time"
 
 	"github.com/Azure/azure-amqp-common-go/auth"
+	"github.com/Azure/azure-amqp-common-go/internal/tracing"
+	"github.com/Azure/azure-amqp-common-go/log"
 	"github.com/Azure/azure-amqp-common-go/rpc"
-	log "github.com/sirupsen/logrus"
 	"pack.ag/amqp"
 )
 
@@ -45,6 +47,9 @@ const (
 
 // NegotiateClaim attempts to put a token to the $cbs management endpoint to negotiate auth for the given audience
 func NegotiateClaim(ctx context.Context, audience string, conn *amqp.Client, provider auth.TokenProvider) error {
+	span, ctx := tracing.StartSpanFromContext(ctx, "az-amqp-common.cbs.NegotiateClaim")
+	span.Finish()
+
 	link, err := rpc.NewLink(conn, cbsAddress)
 	if err != nil {
 		return err
@@ -56,7 +61,7 @@ func NegotiateClaim(ctx context.Context, audience string, conn *amqp.Client, pro
 		return err
 	}
 
-	log.Debugf("negotiating claim for audience %s with token type %s and expiry of %s", audience, token.TokenType, token.Expiry)
+	log.For(ctx).Debug(fmt.Sprintf("negotiating claim for audience %s with token type %s and expiry of %s", audience, token.TokenType, token.Expiry))
 	msg := &amqp.Message{
 		Value: token.Token,
 		ApplicationProperties: map[string]interface{}{
@@ -69,10 +74,10 @@ func NegotiateClaim(ctx context.Context, audience string, conn *amqp.Client, pro
 
 	res, err := link.RetryableRPC(ctx, 3, 1*time.Second, msg)
 	if err != nil {
-		log.Error(err)
+		log.For(ctx).Error(err)
 		return err
 	}
 
-	log.Debugf("negotiated with response code %d and message: %s", res.Code, res.Description)
+	log.For(ctx).Debug(fmt.Sprintf("negotiated with response code %d and message: %s", res.Code, res.Description))
 	return nil
 }